webserver/requirements.txt
python-dotenv
pymongo
python-multipart
fastapi
uvicorn
requests
passlib[bcrypt]
pydantic[email]
pyjwt
httpx
motor
cryptography
rsa
bcrypt==4.0.1
